[LB-461] feat : 트레이싱 설정
